requirements.txt
streamlit
pypdf
langchain
langchain-community
langchain-openai
faiss-cpu
python-dotenv
